// src/api-client/models/InternetConnectionStatuses.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { CicHealthCheckStatus } from "./CicHealthCheckStatus";
import {
  CicHealthCheckStatusFromJSON,
  CicHealthCheckStatusFromJSONTyped,
  CicHealthCheckStatusToJSON,
} from "./CicHealthCheckStatus";

/**
 *
 * @export
 * @interface InternetConnectionStatuses
 */
export interface InternetConnectionStatuses {
  /**
   *
   * @type {CicHealthCheckStatus}
   * @memberof InternetConnectionStatuses
   */
  connectionToWifi: CicHealthCheckStatus;
  /**
   *
   * @type {CicHealthCheckStatus}
   * @memberof InternetConnectionStatuses
   */
  connectionToEthernet: CicHealthCheckStatus;
  /**
   *
   * @type {CicHealthCheckStatus}
   * @memberof InternetConnectionStatuses
   */
  isInternetReachable: CicHealthCheckStatus;
}

/**
 * Check if a given object implements the InternetConnectionStatuses interface.
 */
export function instanceOfInternetConnectionStatuses(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "connectionToWifi" in value;
  isInstance = isInstance && "connectionToEthernet" in value;
  isInstance = isInstance && "isInternetReachable" in value;

  return isInstance;
}

export function InternetConnectionStatusesFromJSON(
  json: any,
): InternetConnectionStatuses {
  return InternetConnectionStatusesFromJSONTyped(json, false);
}

export function InternetConnectionStatusesFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): InternetConnectionStatuses {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    connectionToWifi: CicHealthCheckStatusFromJSON(json["connectionToWifi"]),
    connectionToEthernet: CicHealthCheckStatusFromJSON(
      json["connectionToEthernet"],
    ),
    isInternetReachable: CicHealthCheckStatusFromJSON(
      json["isInternetReachable"],
    ),
  };
}

export function InternetConnectionStatusesToJSON(
  value?: InternetConnectionStatuses | null,
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    connectionToWifi: CicHealthCheckStatusToJSON(value.connectionToWifi),
    connectionToEthernet: CicHealthCheckStatusToJSON(
      value.connectionToEthernet,
    ),
    isInternetReachable: CicHealthCheckStatusToJSON(value.isInternetReachable),
  };
}